requirements/prod.txt
gitappdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4.2
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==3.0.0
cffi==1.11.4
Click==7.0
cmsplugin-filer==1.1.3
coreapi==2.3.0
coreschema==0.0.4
cryptography==2.1.4
diff-match-patch==20121119
dj-database-url==0.4.2
Django==2.0.9
django-appconf==1.0.2
django-braces==1.8.1
django-bulk-update==1.1.10
django-classy-tags==0.8.0
django-cms==3.4.4
django-cors-headers==2.4.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.10.1
django-enumfields==0.8.2
django-filer==1.2.7
django-filter==2.0.0
django-filters==0.2.1
django-formtools==2.0
django-googledrive-storage==1.3.5
django-import-export==1.1.0
django-jet==1.0.7
django-mptt==0.8.7
django-polymorphic==1.0.2
django-queryset-csv==1.0.0
django-rest-swagger==2.1.2
django-sekizai==0.10.0
django-storages==1.4.1
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.6
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
djangorestframework==3.8.2
easy-thumbnails==2.4.1
enum34==1.1.6
et-xmlfile==1.0.1
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
gTTS==2.0.1
gTTS-token==1.1.2
html5lib==0.9999999
httplib2==0.12.0
idna==2.6
isort==4.3.4
itypes==1.1.0
jdcal==1.4
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.10
numpy==1.13.0
oauth2client==4.1.3
odfpy==1.3.6
olefile==0.44
openapi-codec==1.3.1
openpyxl==2.5.8
packaging==16.8
Pillow==4.2.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
pysolr==3.7.0
python-dateutil==2.6.1
pyttsx3==2.7
pytz==2017.2
PyYAML==3.13
requests==2.13.0
rsa==4.0
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.4
tablib==0.12.1
typed-ast==1.1.0
tzlocal==1.4
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.24.1
watson-developer-cloud==1.0.2
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
